\name{cleanLatLon}
\alias{cleanLatLon}
\title{Standardize latitude/longitude coordinates}
\usage{
  cleanLatLon(vec)
}
\arguments{
  \item{vec}{Character vector of lat/long coordinates}
}
\value{
  Numeric decimal lat/lon
}
\description{
  Standardize latitude/longitude coordinates
}

