\name{closestPoint}
\alias{closestPoint}
\title{Find closest point to a given point's coordinates (closestPoint).}
\usage{
closestPoint(point, points)
}
\arguments{
  \item{points}{points is an Nx2 matrix, where columns are
  x,y and rows are the two points}

  \item{point}{the point from which you want to find the
  closest match in points}
}
\value{
Distance, in same units as input
}
\description{
Find closest point to a given point's coordinates
(closestPoint).
}

