\name{cumDist}
\alias{cumDist}
\title{Calculate cumulative distance along a matrix of x,y coordinates}
\usage{
cumDist(coords)
}
\arguments{
  \item{coords}{an [n,2] matrix of coordinates}
}
\value{
A single value that represents the distance along the path
}
\description{
Calculate cumulative distance along a matrix of x,y
coordinates
}
\seealso{
See Also \code{\link{reshapeSLDF}},
\code{\link{SLDFtoLine}}
}

