\name{IDs}
\alias{IDs}
\alias{IDs.SpatialPolygonsDataFrame}
\alias{IDs.default}
\alias{IDs<-}
\alias{IDs<-.SpatialPolygonsDataFrame}
\title{Get sp feature IDs}
\usage{
IDs(x, ...)

\method{IDs}{default}(x, ...)

\method{IDs}{SpatialPolygonsDataFrame}(x, ...)

IDs(x) <- value

IDs(x) <- value
}
\arguments{
  \item{x}{The object to get the IDs from or assign to}

  \item{value}{The character vector to assign to the IDs}

  \item{\dots}{Pass-alongs}
}
\description{
Get sp feature IDs

Assign sp feature IDs
}
\author{
Ari B. Friedman
}

