\name{lineDist}
\alias{lineDist}
\title{Line distance in SpatialLinesDataFrame
Stores length of each line segment in a SpatialLinesDataFrame's data.frame}
\usage{
lineDist(SLDF, varname = "distances")
}
\arguments{
  \item{SLDF}{SpatialLinesDataFrame}

  \item{varname}{Character string containing name of
  variable to hold line distances.}
}
\value{
Returns a SpatialLinesDataFrame
}
\description{
Line distance in SpatialLinesDataFrame Stores length of
each line segment in a SpatialLinesDataFrame's data.frame
}

