\docType{methods}
\name{merge}
\alias{merge}
\alias{merge,SpatialPolygonsDataFrame,data.frame-method}
\title{Merge a SpatialPolygonsDataFrame with a data.frame}
\arguments{
  \item{SPDF}{A SpatialPolygonsDataFrame}

  \item{df}{A data.frame}

  \item{\dots}{Parameters to pass to merge.data.frame}
}
\description{
  Merge a SpatialPolygonsDataFrame with a data.frame
}

