\name{overlayPolyPoly}
\alias{overlayPolyPoly}
\title{Split polygons into contiguous parts
Overlay, in the sense described here: http://resources.esri.com/help/9.3/ArcGISengine/java/Gp_ToolRef/geoprocessing/overlay_analysis.htm}
\usage{
overlayPolyPoly(poly1, poly2)
}
\arguments{
  \item{poly1}{SPDF}

  \item{poly2}{SPDF}
}
\value{
A SPDF with nested polygons of each
}
\description{
Split polygons into contiguous parts Overlay, in the sense
described here:
http://resources.esri.com/help/9.3/ArcGISengine/java/Gp_ToolRef/geoprocessing/overlay_analysis.htm
}

