\name{pointDistPairwise}
\alias{pointDistPairwise}
\title{Create all pairwise distances of points from a SpatialPointsDataFrame}
\usage{
  pointDistPairwise(SPDF, names = "name")
}
\arguments{
  \item{SPDF}{SpatialPointsDataFrame}

  \item{names}{variable name in the SPDF's dataframe used
  to label each point in the resulting matrix}
}
\value{
  matrix
}
\description{
  Create all pairwise distances of points from a
  SpatialPointsDataFrame
}

