\name{smoothLines}
\alias{smoothLines}
\title{Smooth the line segments in a SpatialLinesDataFrame}
\usage{
  smoothLines(lineDF)
}
\arguments{
  \item{lineDF}{SpatialLinesDataFrame}
}
\value{
  SpatialLinesDataFrame
}
\description{
  Smooth the line segments in a SpatialLinesDataFrame.
  Takes a jittery SLDF (such as a GPS produces) and smooths
  it.
}
\seealso{
  See Also as \code{\link{SLDFtoLine}}, ~~~
}

