\name{SPDFtoPointsDF}
\alias{SPDFtoPointsDF}
\title{Convert SpatialPointsDataFrame to a regular data.frame with the coordinates
as "x" and "y"}
\usage{
  SPDFtoPointsDF(SPDF)
}
\arguments{
  \item{SPDF}{SpatialPointsDataFrame}
}
\value{
  data.frame with the coordinates as "x" and "y"
}
\description{
  Convert SpatialPointsDataFrame to a regular data.frame
  with the coordinates as "x" and "y"
}

