\name{subsetSPDF}
\alias{subset.SpatialPointsDataFrame}
\alias{subset.SpatialPolygonsDataFrame}
\alias{subsetSPDF}
\title{Subset SpatialPolygonsDataFrame or SpatialPointsDataFrame}
\usage{
subsetSPDF(x, tf, ...)
}
\arguments{
  \item{x}{SpatialPolygonsDataFrame or
  SpatialPointsDataFrame}

  \item{tf}{Boolean on which to subset}

  \item{\dots}{Arguments to pass to
  SpatialPolygonsDataFrame() or SpatialPointsDataFrame()
  when reconstructing objects}
}
\value{
SpatialPolygonsDataFrame or SpatialPointsDataFrame
}
\description{
Subset SpatialPolygonsDataFrame or SpatialPointsDataFrame
}

