\name{cleanLatLon}
\alias{cleanLatLon}
\title{Standardize latitude/longitude coordinates
Cleans up character representations of lat/lon coordinates}
\usage{
cleanLatLon(vec)
}
\arguments{
  \item{vec}{Character vector of lat/long coordinates}
}
\value{
Numeric decimal lat/lon
}
\description{
Standardize latitude/longitude coordinates Cleans up
character representations of lat/lon coordinates
}

