\name{gGeoCode}
\alias{gGeoCode}
\title{Geocode using Google Maps API (deprecated)
This has been generalized to the taRifx::geocode() function}
\usage{
  gGeoCode(...)
}
\arguments{
  \item{\dots}{Ignored}
}
\description{
  Geocode using Google Maps API (deprecated) This has been
  generalized to the taRifx::geocode() function
}

