\name{IDs}
\alias{IDs}
\alias{IDs.SpatialPolygonsDataFrame}
\alias{IDs.default}
\alias{IDs<-}
\alias{IDs<-.SpatialPolygonsDataFrame}
\title{Get sp feature IDs}
\usage{
  IDs(x, ...)

  \method{IDs}{default} (x, ...)

  \method{IDs}{SpatialPolygonsDataFrame} (x, ...)

  IDs(x, value) <- value

  \method{IDs}{SpatialPolygonsDataFrame} (x, value) <-
    value
}
\arguments{
  \item{x}{The object to get the IDs from}

  \item{\dots}{Pass-alongs}

  \item{x}{The object to assign to}

  \item{value}{The character vector to assign to the IDs}
}
\description{
  Get sp feature IDs

  Assign sp feature IDs
}
\author{
  Ari B. Friedman

  Ari B. Friedman
}

