\docType{methods}
\name{merge,SpatialPolygonsDataFrame,data.frame-method}
\alias{merge,SpatialPolygonsDataFrame,data.frame-method}
\title{Merge a SpatialPolygonsDataFrame with a data.frame}
\usage{
\S4method{merge}{SpatialPolygonsDataFrame,data.frame}(x, y, ...)
}
\arguments{
  \item{SPDF}{A SpatialPolygonsDataFrame}

  \item{df}{A data.frame}

  \item{\dots}{Parameters to pass to merge.data.frame}

  \item{x}{data frames, or objects to be coerced to one.}

  \item{...}{arguments to be passed to or from methods.}
}
\description{
Merge a SpatialPolygonsDataFrame with a data.frame
}

